fix: Remove webpack imports in npm package
