+++ babel.config.js
@@ -18,6 +18,12 @@ module.exports = function (api, ...args) {
     '**/*.component-test.js',
     '**/__mocks__/*.js'
   ]
+  const npmIgnore = [
+    ...ignore,
+    'src/common/config/env.cdn.js',
+    'src/loaders/configure/nonce.cdn.js',
+    'src/loaders/configure/public-path.cdn.js'
+  ]
   const presets = [
     '@babel/preset-env'
   ]
@@ -45,15 +51,7 @@ module.exports = function (api, ...args) {
       ]
     },
     webpack: {
-      ignore,
-      plugins: [
-        [
-          './tools/babel/plugins/transform-import',
-          {
-            '(constants/)env$': '$1env.cdn'
-          }
-        ]
-      ]
+      ignore
     },
     'webpack-ie11': {
       ignore,
@@ -73,53 +71,24 @@ module.exports = function (api, ...args) {
             }
           }
         ]
-      ],
-      plugins: [
-        [
-          './tools/babel/plugins/transform-import',
-          {
-            '(constants/)env$': '$1env.cdn',
-            'polyfill-detection$': 'polyfill-detection.es5'
-          }
-        ]
       ]
     },
     'npm-cjs': {
-      ignore,
+      ignore: npmIgnore,
       presets: [
         [
           '@babel/preset-env', {
             modules: 'commonjs'
           }
         ]
-      ],
-      plugins: [
-        [
-          './tools/babel/plugins/transform-import',
-          {
-            '(/constants/|^\\./)env$': '$1env.npm',
-            '(/configure/|^\\./)public-path$': '$1public-path.npm',
-            '(/configure/|^\\./)nonce$': '$1nonce.npm'
-          }
-        ]
       ]
     },
     'npm-esm': {
-      ignore,
+      ignore: npmIgnore,
       presets: [
         [
           '@babel/preset-env', {
             modules: false
-          }
-        ]
-      ],
-      plugins: [
-        [
-          './tools/babel/plugins/transform-import',
-          {
-            '(/constants/|^\\./)env$': '$1env.npm',
-            '(/configure/|^\\./)public-path$': '$1public-path.npm',
-            '(/configure/|^\\./)nonce$': '$1nonce.npm'
           }
         ]
       ]
