+++ src/common/constants/env.js
@@ -1,28 +1,26 @@
 /**
- * @file This file exposes build time environment variables that will be included in the
- * build output of the agent. This file specifically contains the normal environment variables
- * for the NPM agent build and will be overridden by webpack/babel during the build based on the
- * type of build being performed.
+ * @file This file exposes NPM build environment variables. These variables will
+ * be overridden with babel.
  */
-
-import pkgJSON from '../../../package.json'
 
 /**
  * Exposes the version of the agent
  */
-export const VERSION = pkgJSON.version
+export const VERSION = process.env.BUILD_VERSION
 
 /**
  * Exposes the build type of the agent
+ * Valid values are LOCAL, PROD, DEV
  */
 export const BUILD_ENV = 'NPM'
 
 /**
  * Exposes the distribution method of the agent
+ * Valid valuse are CDN, NPM
  */
 export const DIST_METHOD = 'NPM'
 
 /**
  * Exposes the lib version of rrweb
  */
-export const RRWEB_VERSION = pkgJSON.dependencies.rrweb
+export const RRWEB_VERSION = process.env.RRWEB_VERSION
