--- src/loaders/configure/nonce.npm.js
@@ -1,1 +1,0 @@
-// We don't support setting automating the nonce attribute in the npm package
