+++ src/loaders/configure/public-path.js
@@ -1,9 +1,3 @@
-// Set the default CDN or remote for fetching the assets; NPM shouldn't change this var.
-
-export const redefinePublicPath = (urlString) => {
-  const isOrigin = urlString.startsWith('http')
-  // Input is not expected to end in a slash, but webpack concats as-is, so one is inserted.
-  urlString += '/'
-  // If there's no existing HTTP scheme, the secure protocol is prepended by default.
-  __webpack_public_path__ = isOrigin ? urlString : 'https://' + urlString // eslint-disable-line
+export const redefinePublicPath = () => {
+  // We don't support setting public path in webpack via NPM build.
 }
