--- src/loaders/configure/public-path.npm.js
@@ -1,3 +1,0 @@
-export const redefinePublicPath = () => {
-  // We don't support setting public path in webpack via NPM build.
-}
