+++ tsconfig.json
@@ -10,5 +10,12 @@
     "skipLibCheck": true
   },
   "include": ["src/**/*"],
-  "exclude": ["**/*.test.js", "**/*.component-test.js","**/__mocks__/*.js"]
+  "exclude": [
+    "**/*.test.js",
+    "**/*.component-test.js",
+    "**/__mocks__/*.js",
+    "src/common/config/env.cdn.js",
+    "src/loaders/configure/nonce.cdn.js",
+    "src/loaders/configure/public-path.cdn.js"
+  ]
 }
